+++ src/Bookings.php
@@ -56,7 +56,7 @@ class Bookings extends Plugin
 	// Properties
 	// =========================================================================
 
-	public $schemaVersion = '1.0.3';
+	public $schemaVersion = '1.0.4';
 
 	public $hasCpSettings = true;